refactor: expose xz binary path via flags
The binary path of xz, and its use, is currently magic computed within
the xz package; This is not a very nice UX. Move it to flags set on the
main command so the user can enable it, disable it or set a custom xz
binary path.
