+++ pkg/xz/export_test.go
@@ -2,8 +2,15 @@ package xz
 
 import (
 	"context"
+	"errors"
 	"io"
+	"os"
+	"os/exec"
+	"path/filepath"
 )
+
+// ErrXZBinAbsPath is returned when XZ_BINARY_PATH is not an absolute path.
+var ErrXZBinAbsPath = errors.New("XZ_BINARY_PATH must be an absolute path")
 
 func DecompressCommand(ctx context.Context, r io.Reader) (io.ReadCloser, error) {
 	p, err := getXZBin()
@@ -16,4 +23,16 @@ func DecompressCommand(ctx context.Context, r io.Reader) (io.ReadCloser, error) 
 
 func DecompressInternal(ctx context.Context, r io.Reader) (io.ReadCloser, error) {
 	return decompressInternal(ctx, r)
+}
+
+func getXZBin() (string, error) {
+	if p := os.Getenv("XZ_BINARY_PATH"); p != "" {
+		if !filepath.IsAbs(p) {
+			return "", ErrXZBinAbsPath
+		}
+
+		return p, nil
+	}
+
+	return exec.LookPath("xz")
 }
